fix(run dev): Fixed the "npm run dev" problem
As in #385, #421 and #425, I was able to make this repo work out for me
doing this steps:
```md
I'm still learning VueJS and trying to improve my skills. Yesterday I was able to figure out (after search on the web) on how to make `npm run dev` works.

I've done this steps:

1. ### NPM Update
`npm update`

2.  ### NPM install webpack-merge
`npm i webpack-merge`

3.  ### Edit package json
```json
"scripts":{
...
	"dev": "NODE_OPTIONS='--openssl-legacy-provider' node build/dev-server.js"
...
}
```

After that, I was able to run the server on localhost.
```